pyproject.toml
[tool.poetry]
name = "plantotron"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
arduino-simple-rpc = "^2.4.2"
click = "^8.1.3"
google-api-python-client = "^2.88.0"
simple-pid = "^2.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
